+++ enum_string_extension_gen/example/example.dart
@@ -4,8 +4,7 @@ import 'package:enum_string_extension/enum_string_extension.dart';
 import 'package:flutter/cupertino.dart' show Localizations, BuildContext;
 import 'package:meta/meta.dart' show immutable;
 
-/// Make sure that `part` is specified, even before launching the builder
-part 'example.g.dart';
+part 'basic_class.g.dart';
 
 // The localizations delagate must be called AppLocalizations and implement each field of the enum
 class AppLocalizations {
@@ -16,6 +15,7 @@ class AppLocalizations {
     const Locale.fromSubtags(languageCode: 'en', countryCode: 'US'): const AppLocalizationsLabels(
       value1: 'Value 1 in english',
       value2: 'Value 2 in english',
+      value3: 'Value 3 in english',
     )
   };
 
@@ -25,17 +25,44 @@ class AppLocalizations {
       Localizations.of<AppLocalizations>(context, AppLocalizations)?.labels;
 }
 
+class AppLocalizationsLabels {
+  const AppLocalizationsLabels({
+    this.value1,
+    this.value2,
+    this.value3,
+  });
+
+  final String value1;
+  final String value2;
+  final String value3;
+}
+
+enum TestEnum3 {
+  value1,
+  value3,
+}
+
+enum TestEnum2 {
+  value1,
+  value2,
+  value3,
+}
+
 enum TestEnum {
   value1,
   value2,
 }
 
 @immutable
-// Annotation cannot be on the enum directly, one of the class containing the enum and with the enum declaration in
-// the same file must be annotated with the `@enumString` annotation.
 @enumString
-class SimpleObject {
-  final TestEnum value;
+class BasicClass {
+  const BasicClass({
+    this.test,
+    this.test2,
+    this.test3,
+  });
 
-  const SimpleObject({this.value});
+  final TestEnum test;
+  final TestEnum2 test2;
+  final List<TestEnum3> test3;
 }