+++ enum_string_extension_gen/test/constants.dart
@@ -32,6 +32,7 @@ class AppLocalizations {
     const Locale.fromSubtags(languageCode: 'en', countryCode: 'US'): const AppLocalizationsLabels(
       value1: 'Value 1 in english',
       value2: 'Value 2 in english',
+      value3: 'Value 3 in english',
     )
   };
 
@@ -42,10 +43,26 @@ class AppLocalizations {
 }
 
 class AppLocalizationsLabels {
-  const AppLocalizationsLabels({this.value1, this.value2});
+  const AppLocalizationsLabels({
+    this.value1,
+    this.value2,
+    this.value3,
+  });
 
   final String value1;
   final String value2;
+  final String value3;
+}
+
+enum TestEnum3 {
+  value1,
+  value3,
+}
+
+enum TestEnum2 {
+  value1,
+  value2,
+  value3,
 }
 
 enum TestEnum {
@@ -56,9 +73,15 @@ enum TestEnum {
 @immutable
 @enumString
 class BasicClass {
-  const BasicClass({this.test});
+  const BasicClass({
+    this.test,
+    this.test2,
+    this.test3,
+  });
 
   final TestEnum test;
+  final TestEnum2 test2;
+  final List<TestEnum3> test3;
 }
 ''';
 
@@ -80,6 +103,38 @@ extension TestEnumStringExtension on TestEnum {
         return AppLocalizations.of(context).value1;
       case TestEnum.value2:
         return AppLocalizations.of(context).value2;
+
+      default:
+        break;
+    }
+    return AppLocalizations.of(context).value1;
+  }
+}
+
+extension TestEnum2StringExtension on TestEnum2 {
+  String text(BuildContext context) {
+    switch (this) {
+      case TestEnum2.value1:
+        return AppLocalizations.of(context).value1;
+      case TestEnum2.value2:
+        return AppLocalizations.of(context).value2;
+      case TestEnum2.value3:
+        return AppLocalizations.of(context).value3;
+
+      default:
+        break;
+    }
+    return AppLocalizations.of(context).value1;
+  }
+}
+
+extension TestEnum3StringExtension on TestEnum3 {
+  String text(BuildContext context) {
+    switch (this) {
+      case TestEnum3.value1:
+        return AppLocalizations.of(context).value1;
+      case TestEnum3.value3:
+        return AppLocalizations.of(context).value3;
 
       default:
         break;