+++ enum_string_extension_test/lib/basic_class.g.dart
@@ -22,3 +22,35 @@ extension TestEnumStringExtension on TestEnum {
     return AppLocalizations.of(context).value1;
   }
 }
+
+extension TestEnum2StringExtension on TestEnum2 {
+  String text(BuildContext context) {
+    switch (this) {
+      case TestEnum2.value1:
+        return AppLocalizations.of(context).value1;
+      case TestEnum2.value2:
+        return AppLocalizations.of(context).value2;
+      case TestEnum2.value3:
+        return AppLocalizations.of(context).value3;
+
+      default:
+        break;
+    }
+    return AppLocalizations.of(context).value1;
+  }
+}
+
+extension TestEnum3StringExtension on TestEnum3 {
+  String text(BuildContext context) {
+    switch (this) {
+      case TestEnum3.value1:
+        return AppLocalizations.of(context).value1;
+      case TestEnum3.value3:
+        return AppLocalizations.of(context).value3;
+
+      default:
+        break;
+    }
+    return AppLocalizations.of(context).value1;
+  }
+}