+++ enum_string_extension/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2 Fix multiple enum and support subtypes for `List<T>`
+
+* Fix bug that prevented multiple enum to generate code
+* Add support for types inside lists
+
 ## 0.0.1 Initial release
 
-* Lets you annotate an object using `@enumString` to generate a `enumString` extension.
+* Lets you annotate an object using `@enumString` to generate a `enumString` extension